// web/src/generated/api/types.gen.ts
export type TaskSummaryResponse = {
    /**
     * Adapter
     */
    adapter: string;
    /**
     * Display Name
     */
    display_name: string;
    /**
     * Doc Path
     */
    doc_path?: string | null;
    /**
     * Has Evaluator
     */
    has_evaluator: boolean;
    /**
     * Has Init
     */
    has_init: boolean;
    /**
     * Is Stub
     */
    is_stub: boolean;
    /**
     * Problem Summary
     */
    problem_summary?: string;
    /**
     * Task Id
     */
    task_id: string;
};